+++ cookbooks/arcgis-desktop/CHANGELOG.md
@@ -3,6 +3,11 @@ arcgis-desktop cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-desktop cookbook.
 
+3.4.0
+-----
+- Added support for ArcGIS 10.7.1.
+- Updated and locked versions of depenent cookbooks.
+
 3.3.0
 -----
 - Added support for ArcGIS 10.7.