+++ cookbooks/arcgis-enterprise/CHANGELOG.md
@@ -3,10 +3,10 @@ arcgis-enterprise cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-enterprise cookbook.
 
-3.3.1
+3.4.0
 -----
-- Workaround for [BUG-000121142](https://support.esri.com/en/technical-article/000020633)
-- Increased configuredatastore tool execution timeout to 3 hours.
+- Added support for ArcGIS Enterprise 10.7.1.
+- Updated and locked versions of depenent cookbooks.
 
 3.3.0
 -----